+++ core/src/main/java/com/graphhopper/routing/util/parsers/OSMRoadClassParser.java
@@ -49,6 +49,10 @@ public class OSMRoadClassParser implements TagParser {
         String roadClassTag = readerWay.getTag("highway");
         if (roadClassTag == null)
             return edgeFlags;
+        if (!roadClassTag.isEmpty() && readerWay.hasTag("motorroad", "yes")) {
+            roadClassEnc.setEnum(false, edgeFlags, RoadClass.MOTORWAY);
+            return edgeFlags;
+        }
         RoadClass roadClass = RoadClass.find(roadClassTag);
         if (roadClass == OTHER && roadClassTag.endsWith("_link"))
             roadClass = RoadClass.find(roadClassTag.substring(0, roadClassTag.length() - 5));